feat(inference): curated model list, closed device and precision fields, download lifecycle (#470)
* feat(kernel): device and precision are closed vocabularies, and weights can be verified

Two rules the kernel could state and did not.

`device` accepts `cpu`, `cuda` and `cuda:N` and nothing else. Everything
outside that used to be accepted and then quietly resolved onto the CPU in
full precision, so a connection saying `gpu` described a run that never
happened and went on displaying `gpu` while it did not happen.

`precision` is a `Precision` enum — `fp16`, `fp32` — with the spellings this
build already honoured (`float16`, `half`, `float32`, `full`) normalized onto
it, so a row written under the free-text field stays readable. It is
conditioned on the device: both local adapters resolve half precision as *this
device is CUDA and the connection asked for fp16*, which makes `cpu` + `fp16`
a setting with no effect rather than a slow one. `precisions_for` owns that
rule so a form offering a choice and the kernel refusing one read the same
function.

`download_weights` becomes legal at `{local, ready}` as well. The work is
idempotent by the download library's own design — a snapshot already in the
cache is checked against its hashes rather than re-fetched — so the same call
against a ready connection answers "are the weights still there and still
intact?", which had no action at all before. That retires the `retrying` flag:
it existed only to relax the state half of a gate that no longer refuses, and
a re-queued orphan now takes the ordinary path.

cf. #469

* feat(app): a curated model list, closed device and precision fields, and a download that finishes on screen

The connection form asked a person to type three things it could have offered
them. All three are now lists.

**Model.** A grouped select over the models this build has an adapter for, each
entry pinned to a revision verified against the locked transformers and shown
with the hub's own download size and a line on what it is for. **Custom model…**
reveals the free model id and revision fields, so curation guides without
restricting. The list lives in one module, `inferenceCatalog.ts`, and the form
reads it and nothing else — a model id spelled anywhere in `screens/` is now a
test failure.

**Device and precision.** The kernel's vocabularies, offering-side, with the
precision list following the device because half precision applies on CUDA only.
A row already carrying `cuda:1` keeps it rather than being silently reassigned.
The kernel is still the authority: what it refuses arrives as prose.

**The download's whole life.** A settled job invalidates the connection list, so
a finished download turns the row `Ready` where somebody can see it — the bug
was that only the `202` invalidated, and the completion, which is what moves
`setup_state`, invalidated nothing. A failure renders the job's own words plus
what to do about it, the row stays `Not set up`, and **Download weights** is the
retry — there is no second control. At `Ready` the same declared action appears
as **Verify weights** in the row's overflow, where it re-checks that the
snapshot is complete.

The claim that a re-run verifies the cache against its hashes was checked and is
not true: `hf_hub_download` returns a file already present under a pinned
revision without re-reading it. What a re-run establishes is completeness, and
the docstrings, the docs and the UI copy now say that instead.

cf. #469

+++ tests/kernel/test_metadata_store.py
@@ -194,7 +194,7 @@ def _seed(uow: UnitOfWork) -> list[tuple[str, UUID]]:
             model_id="some/model",
             model_revision="abc123",
             device="cpu",
-            precision="fp16",
+            precision="fp32",
             created_at=datetime(2026, 8, 7, 12, 0, tzinfo=UTC),
             updated_at=datetime(2026, 8, 7, 12, 0, tzinfo=UTC),
         )
